Penmqtt/requirements.txt
paho-mqtt==1.6.1
python-nmap==0.7.1
reportlab==4.1.0
scapy==2.5.0
netifaces==0.11.0
colorama==0.4.6
prettytable==3.10.0
PySide6
